Reorganised data to reduce sideEffects
